requirements.txt
streamlit
langchain-openai
azure.identity